Fixed ATP 02-02-003

+++ remove_incredibuild_package.sh
@@ -78,11 +78,15 @@ function remove_scripts() {
 
 function remove_ssh_addon() {
     echo "Removing ssh addon files:"
-    for i in $SSH_ADDONS;
-    do
-        echo "removing $i";
-        rm -vfr $i;
-    done
+    if [[ -e /bin/GridCoordinator ]]; then
+        echo "ssh is not removing, it shared with Coordinator"
+    else
+        for i in $SSH_ADDONS;
+        do
+            echo "removing $i";
+            rm -vfr $i;
+        done
+    fi
 }
 
 function remove_user() {
